Improve specs with a custom matcher
Although the raw LOC seems to be increased and verbosity is much
greater, overall the spec is "cleaner" to understand and follow along
now. Plus the messages in `doc` output are easier to understand.

This will also be helpful when more functionality starts being added and
we want to potentially use this matcher across other specs.