Ensure `Module.submodule` uses a false `inherit` flag
The `const_get` function accepts an `inherit` flag which is true by
default, meaning it will search ancestors for a valid constant. This
might be good in some situations, but it contradicts the function
statement, which is purely a _sub_module, meaning it should directly
descend from the calling module and not any of the parents.