Improve source code inspection logic and readability
One of the biggest logic changes is matching on the guessed name
completey. Before we were saying "if the file name matches anywhere in
the captured class or module name, use it" which was too broad.

Take the example code:

```ruby
module Subway
  class Sub
    ... stuff ...
  end
end
```

The previous logic would match the file name `sub` with `module Subway`
and use that, which wouldn't be accurate. So anything where the file
name was an exact subset of characters in another class/module name that
came before the desired one, you could incorrectly match and get some
errors.