Add `exclude` option to autoloader
Takes in an array of strings and regexps. The current implementation
only checks against the name of the module/class, however, so there's no
way to specify based on a path. There's also no way to exlucde based on
a namespace yet, unless you want to exclude an entire namespace.