Use whitequark's ast and parser gems to perform more intelligent loading
The Parser and AST libraries allow us to "fake load" a Ruby source file
and determine which modules and classes are going to attempt to load.
Although not perfect (yet), it allows us to remove some of the options
that were being used to supplement the previous method of text scanning.

Another benefit is the uncoupling of namespaces from directory
structure. Previously, ACLS required a fairly strict directory structure
because it made assumptions between that structure and the construction
of module namespaces. Using Parser and AST, it is no longer necessary to
make any such assumptions as we can reliably parse out and interpret
source files in regards to which module namespaces are being
constructed.

Still requires work to make the whole thing even smarter, but it's a
significant jump forward (I believe) from the previous version.